reorganized bootstrapping and paths
