+++ src/index.js
@@ -1,13 +1,9 @@
-import componentManager from 'lib/core/ComponentManager'
-import Component from 'brindille-component'
-import { initRouter } from 'lib/router'
+import brindille from './lib/brindille'
+import { initRouter } from './lib/router'
 import * as views from './views'
 import './styles/index.styl'
 
-componentManager.registerMultiple(views)
+const app = brindille(document.body, views)
 
-let rootComponent = new Component(document.body, componentManager.get)
-componentManager.setRootComponent(rootComponent)
-
-const router = initRouter(rootComponent)
+const router = initRouter(app)
 router.start()
