+++ src/views/sections/about/About.js
@@ -1,3 +1,3 @@
-import Section from 'lib/core/Section'
+import Section from '../../../lib/Section'
 
 export default class About extends Section {}
