+++ src/views/sections/home/Home.js
@@ -1,3 +1,3 @@
-import Section from 'lib/core/Section'
+import Section from '../../../lib/Section'
 
 export default class Home extends Section {}
