+++ src/views/sections/post/Post.js
@@ -1,3 +1,3 @@
-import Section from 'lib/core/Section'
+import Section from '../../../lib/Section'
 
 export default class Post extends Section {}
